+++ adapters/Line/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## LINE iOS Mediation Adapter Changelog
 
+#### Next version
+- Removed logic that checked whether the FAD SDK was registered in the adapter setup.
+
 #### [Version 2.9.20241106.1](https://dl.google.com/googleadmobadssdk/mediation/ios/line/LineAdapter-2.9.20241106.1.zip)
 - Updated the adapter to use the latest ad load API for bidding banner ads.
 - Fixed an issue preventing ad events from forwarding correctly for bidding interstitial ads.