+++ adapters/Line/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updated the interstitial and rewarded ads' "show" API with new "showWithViewController:" API.
 - Removed native ad state check before downloading native ad assets, because it is deprecated and there is no replacement.
 - Fixed how test mode is set on the LINE SDK.
+- Now requires Google Mobile Ads SDK version 12.0.0 or higher.
 
 #### [Version 2.9.20241106.1](https://dl.google.com/googleadmobadssdk/mediation/ios/line/LineAdapter-2.9.20241106.1.zip)
 - Updated the adapter to use the latest ad load API for bidding banner ads.