+++ adapters/Line/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next Version
 - Fixed an issue where `GADMediationAdapterLine.h` was not a public header.
+- Enable -fobjc-arc and -fstack-protector-all.
 
 #### [Version 2.9.20241106.2](https://dl.google.com/googleadmobadssdk/mediation/ios/line/LineAdapter-2.9.20241106.2.zip)
 - Added the `GADMediationAdapterLine.testMode` property to indicate whether the FiveAd SDK should be initialized in test mode. This flag must be set before initializing the Google Mobile Ads SDK.